Bug fixes for navs

+++ src/Structures/NavTreeRepository.php
@@ -3,6 +3,7 @@
 namespace Statamic\Eloquent\Structures;
 
 use Statamic\Contracts\Structures\Tree as TreeContract;
+use Statamic\Eloquent\Structures\NavTree;
 use Statamic\Facades\Blink;
 use Statamic\Stache\Repositories\NavTreeRepository as StacheRepository;
 
@@ -22,6 +23,11 @@ class NavTreeRepository extends StacheRepository
 
     public function save($entry)
     {
+        // if we are using flat files for the config, but eloquent for the data
+        if (! $entry instanceof NavTree) {
+            return parent::save($entry);
+        }
+
         $model = $entry->toModel();
         $model->save();
 
@@ -32,6 +38,11 @@ class NavTreeRepository extends StacheRepository
 
     public function delete($entry)
     {
+        // if we are using flat files for the config, but eloquent for the data
+        if (! $entry instanceof NavTree) {
+            return parent::save($entry);
+        }
+
         $entry->model()->delete();
     }
 }
